requirements.txt
astroid==2.4.2
brotlipy==0.7.0
certifi==2020.12.5
cffi
click==7.1.2
cryptography
idna
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
ortools
protobuf==3.13.0
pycparser
pylint==2.6.0
pyOpenSSL
PySocks
python-dotenv
pywin32==300
selenium
six==1.15.0
toml==0.10.2
urllib3
win-inet-pton
wincertstore==0.2
wrapt==1.12.1
